fix error on retur pembelian

+++ pages/dashboard/pembelian/utility/createInventory.js
@@ -20,7 +20,7 @@ async function createInventory(row) {
     console.log("element purchasing detail", element);
 
     const unitOrder = element.attributes.unit_order;
-    const totalOrder = parseInt(element.attributes.total_order);
+    const totalOrder = parseFloat(element.attributes.total_order);
     const product = element.attributes.product.data;
     const location = element.attributes.location.data;
     const expDate = element.attributes.expired_date;
@@ -52,14 +52,17 @@ async function createInventory(row) {
 }
 
 async function addToGudang(body) {
-  const response = await fetch(`${process.env.NEXT_PUBLIC_URL}/inventories/add`, {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-      Authorization: `Bearer ${cookies.token}`,
-    },
-    body: JSON.stringify(body),
-  });
+  const response = await fetch(
+    `${process.env.NEXT_PUBLIC_URL}/inventories/add`,
+    {
+      method: "POST",
+      headers: {
+        "Content-Type": "application/json",
+        Authorization: `Bearer ${cookies.token}`,
+      },
+      body: JSON.stringify(body),
+    }
+  );
 
   const data = await response.json();
