+++ pages/dashboard/penjualan/toko/pembayaran/retur.js
@@ -346,8 +346,8 @@ function PembayaranToko({ props }) {
                         dataPayment.length > 1
                           ? dataPayment.reduce(
                               (acc, curr) =>
-                                parseInt(acc) +
-                                parseInt(curr.attributes.payment),
+                                parseFloat(acc) +
+                                parseFloat(curr.attributes.payment),
                               0
                             )
                           : dataPayment.length === 1
@@ -382,7 +382,8 @@ function PembayaranToko({ props }) {
                       if (record.attributes.status === "Dibayar") {
                         const dataPaymentValue = dataPayment.reduce(
                           (acc, curr) =>
-                            parseInt(acc) + parseInt(curr.attributes.payment),
+                            parseFloat(acc) +
+                            parseFloat(curr.attributes.payment),
                           0
                         );
                         const kembali =
